MAINT: Moving Mask to the generic module.
